Now supports floats.
This was a simple change to make, maybe at some point later I'll add some functions into the parser.

+++ src/com/Math/Node.java
@@ -8,5 +8,5 @@ public interface Node {
   Node RightNode = null;
   ArrayList<Node> GetNodes();
   public String toString();
-  int QuickParse();
+  float QuickParse();
 }
